requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.8.2
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.2
prefect==2.20.8
python-dotenv==1.0.1
joblib==1.4.2
jinja2==3.1.4
matplotlib==3.9.2
flake8==7.1.1
black==24.8.0
pytest==8.3.3
httpx==0.27.2
mlxtend==0.23.1
prometheus-client
deepchecks
setuptools
